Issue #21 - RF07 e Issue 24 - RF10
Close #21
Close #24

+++ hurri-games-app/database/migrations/2022_12_12_044721_create_reviews_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained();
             $table->foreignId('game_id')->constrained();
             $table->string('review');
-            $table->decimal('rate');
+            $table->integer('rate');
 
             $table->timestamps();
         });
